Factor out table and view lists
