+++ webapp/src/js/components/containers/StartTab.js
@@ -1,15 +1,12 @@
-import _keys from 'lodash/keys';
-import _map from 'lodash/map';
 import React from 'react';
 
 import PureRenderMixin from 'mixins/PureRenderMixin';
-import ConfigMixin from 'mixins/ConfigMixin';
 import FluxMixin from 'mixins/FluxMixin';
+import ConfigMixin from 'mixins/ConfigMixin';
 
 // Material UI
-import List from 'material-ui/lib/lists/list';
-import ListItem from 'material-ui/lib/lists/list-item';
-import Subheader from 'material-ui/lib/Subheader';
+import ViewList from 'panoptes/ViewList';
+import TableList from 'panoptes/TableList';
 
 import Icon from 'ui/Icon';
 
@@ -17,11 +14,11 @@ import HTMLWithComponents from 'panoptes/HTMLWithComponents';
 
 import 'start-tab.scss';
 
-let EmptyTab = React.createClass({
+let StartTab = React.createClass({
   mixins: [
     PureRenderMixin,
-    ConfigMixin,
-    FluxMixin
+    FluxMixin,
+    ConfigMixin
   ],
 
   propTypes: {
@@ -35,47 +32,27 @@ let EmptyTab = React.createClass({
     return 'Start';
   },
 
-  handleOpen(e, container, props) {
-    let actions = this.getFlux().actions.session;
-    actions.tabOpen(container, props, true);
-  },
-
-  handleTableClick(e, table) {
-    if (this.config.tables[table.id].settings.listView) {
-      this.handleOpen(e, 'containers/ListWithActions', {table: table.id});
-    } else {
-      this.handleOpen(e, 'containers/DataTableWithActions', {table: table.id});
+  handleClick({container, props, middleClick}) {
+    if (middleClick)
+      this.flux.actions.tabOpen(container, props, false);
+    else {
+      this.props.componentUpdate(props, container);
     }
   },
 
   render() {
-    let {tables, chromosomes} = this.config;
     return (
-      <div className="horizontal stack start-align wrap">
-
-          <HTMLWithComponents className="grow description">
-            {this.config.settings.description}
-          </HTMLWithComponents>
-        <List style={{width: '400px'}}>
-          <Subheader>Open a view:</Subheader>
-          <ListItem primaryText="Genome Browser"
-                    secondaryText="View table data and sequence data on the genome"
-                    secondaryTextLines={2}
-                    leftIcon={<div><Icon fixedWidth={true} name="bitmap:genomebrowser.png"/></div>}
-                    onClick={(e) => this.handleOpen(e, 'containers/GenomeBrowserWithActions', {chromosome: _keys(chromosomes)[0]})} />
-          <Subheader>Open a table:</Subheader>
-          {_map(tables, (table) => (
-            <ListItem key={table.id}
-                      primaryText={table.tableCapNamePlural}
-                      secondaryText={table.description}
-                      secondaryTextLines={2}
-                      leftIcon={<div><Icon fixedWidth={true} name={table.icon}/></div>}
-                      onClick={(e) => this.handleTableClick(e, table)} />
-          ))}
-        </List>
+      <div className="horizontal stack start-align">
+        <HTMLWithComponents className="grow description">
+          {this.config.settings.description}
+        </HTMLWithComponents>
+        <div className="" >
+          <ViewList style={{width: '400px'}} onClick={this.handleClick} />
+          <TableList style={{width: '400px'}} onClick={this.handleClick} />
+        </div>
       </div>
     );
   }
 });
 
-module.exports = EmptyTab;
+module.exports = StartTab;
