+++ webapp/src/js/components/panoptes/TableList.js
@@ -1,0 +1,55 @@
+import React from 'react';
+import _map from 'lodash/map';
+
+import PureRenderMixin from 'mixins/PureRenderMixin';
+import ConfigMixin from 'mixins/ConfigMixin';
+
+import List from 'material-ui/lib/lists/list';
+import ListItem from 'material-ui/lib/lists/list-item';
+import Subheader from 'material-ui/lib/Subheader';
+import Icon from 'ui/Icon';
+
+let EmptyTab = React.createClass({
+  mixins: [
+    PureRenderMixin,
+    ConfigMixin
+  ],
+
+  propTypes: {
+    style: React.PropTypes.object,
+    onClick: React.PropTypes.func.isRequired
+  },
+
+
+  handleOpen(e, container, props) {
+    const middleClick =  e.button == 1 || e.metaKey || e.ctrlKey;
+    this.props.onClick({container, props, middleClick});
+  },
+
+  handleTableClick(e, table) {
+    if (this.config.tables[table.id].settings.listView) {
+      this.handleOpen(e, 'containers/ListWithActions', {table: table.id});
+    } else {
+      this.handleOpen(e, 'containers/DataTableWithActions', {table: table.id});
+    }
+  },
+
+  render() {
+    let {tables} = this.config;
+    return (
+      <List style={this.props.style}>
+        <Subheader>Open a table:</Subheader>
+        {_map(tables, (table) => (
+          <ListItem key={table.id}
+                    primaryText={table.tableCapNamePlural}
+                    secondaryText={table.description}
+                    leftIcon={<div><Icon fixedWidth={true} name={table.icon}/></div>}
+                    onClick={(e) => this.handleTableClick(e, table)} />
+        ))}
+      </List>
+
+    );
+  }
+});
+
+module.exports = EmptyTab;
